+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # CHANGELOG minishell by nburchha & fschuber
 
-## v0.0.1 - 07.02.24
+## v0.1.0 - 07.02.24
+- added `printing.c` for printing debug functions (with print_tokens())
+- added `testing.c` for testing debug functions (with test_lexer())
+- added non-finished `lexer.c` that splits up input string into token array
+- added deprecated folder with alternative token space seperator approach for unused but good code of potential future use
+
+### v0.0.2 - 07.02.24
 - Makefile spelling fix & executable placement error
 	- added normcomp target to allow for comp style execution without norm checks
+
+### v0.0.1 - 30.01.24
+- readme line adjustments
 
 ## v0.0.0 - 30.01.24
 - Basic Project Setup